@@ -1,12 +1,12 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     repositories {
-        maven { url 'https://maven.google.com' }
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.1.51"
+        classpath 'com.android.tools.build:gradle:3.3.2'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.21"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -15,7 +15,7 @@ buildscript {
 
 allprojects {
     repositories {
-        maven { url 'https://maven.google.com' }
+        google()
         jcenter()
         mavenCentral()
     }
@@ -28,14 +28,12 @@ task clean(type: Delete) {
 ext {
     // sdk tools
     minSdkVersion = 15
-    compileSdkVersion = 27
-    targetSdkVersion = 27
-    buildToolsVersion = "26.0.2"
+    compileSdkVersion = 28
+    targetSdkVersion = 28
 
     // libraries
-    supportLibraryVersion = '27.0.2'
-    kotlinVersion = '1.1.4-3'
-    butterKnifeVersion = '8.6.0'
+    supportLibraryVersion = '1.0.0-beta01'
+    kotlinVersion = '1.3.21'
     retrofitVersion = '2.3.0'
     okhttpVersion = '3.9.0'
     glideVersion = '4.0.0-RC1'
