@@ -1,4 +1,4 @@
-package com.github.ramonrabello.kiphy.data
+package com.github.ramonrabello.kiphy.common.extensions
 
 import android.view.View
 
@@ -13,20 +13,20 @@ fun View.gone() = run { this.visibility = View.GONE }
  * attribute value of type 'tag1-tag2-tag3-<slug_hash>' into
  * '#tag1 #tag2 #tag3' format.
  */
-fun String.tagfy() : String {
+fun String.slugfy() : String {
     var slug = ""
 
     // splits words using delimiter '-'
     this.split("-")
 
             // map each list element to include '#' as first char
-            .map{ it -> "#$it " }
+            .map{ "#$it " }
 
             // ignore the last occurrence in list which is the slug hash
             .dropLast(1)
 
             // concat all elements and save in a final String
-            .forEach { it ->  slug += it }
+            .forEach { slug += it }
 
     return slug.trim()
 }
