@@ -1,13 +1,13 @@
-package com.github.ramonrabello.kiphy.presentation.trends
+package com.github.ramonrabello.kiphy.trends
 
 import android.graphics.Rect
-import android.support.v7.widget.RecyclerView
 import android.view.View
+import androidx.recyclerview.widget.RecyclerView
 
 /**
  * Spacing item decoration for recycler view.
  */
-class SpacingItemDecoration(val itemOffset:Int) : RecyclerView.ItemDecoration() {
+class SpacingItemDecoration(private val itemOffset:Int) : RecyclerView.ItemDecoration() {
 
     override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
         outRect.apply {
